feat: misc RE
