+++ include/RE/H/HUDData.h
@@ -2,9 +2,13 @@
 
 #include "RE/B/BSString.h"
 #include "RE/I/IUIMessageData.h"
+#include "RE/M/MARKER_TYPE.h"
 
 namespace RE
 {
+	class TESQuest;
+	class TESWordOfPower;
+
 	class HUDData : public IUIMessageData
 	{
 	public:
@@ -31,37 +35,29 @@ namespace RE
 			kObjectiveComplete = 18,
 			kObjectiveFailed = 19,
 			kSkillIncrease = 20,
-			kWordOfPower1 = 21,
-			kWordOfPower2 = 22,
+			kWordOfPowerLearned = 21,
+			kDragonSoulAbsorbed = 22,
 			kCrosshairHide = 23,
 			kCrosshairSneak = 24,
-		};
-
-		enum class Discovery
-		{
-			kCity = 1,
-			kTown = 2,
-			kDungeon4 = 4,
-			kDungeon7 = 7,
-			kDungeon8 = 8,
-			kDungeon12 = 12,
-			kDungeon15 = 15
+			kLocationDiscovery = 26,
+			kShowLocationName = 29,
+			kShowHintText = 30,
 		};
 
 		~HUDData() override;  // 00
 
 		// members
-		REX::EnumSet<Type, std::uint32_t>      type;          // 10
-		std::uint32_t                          pad14;         // 14
-		BSString                               text;          // 18
-		ObjectRefHandle                        crossHairRef;  // 28
-		std::uint32_t                          pad2C;         // 2C
-		void*                                  unk30;         // 30
-		std::uint64_t                          unk38;         // 38
-		std::uint8_t                           unk40;         // 40
-		std::uint8_t                           pad41;         // 41
-		std::uint16_t                          pad42;         // 42
-		REX::EnumSet<Discovery, std::uint32_t> discovery;     // 44
+		REX::EnumSet<Type, std::uint32_t>        type;          // 10
+		std::uint32_t                            pad14;         // 14
+		BSString                                 text;          // 18
+		ObjectRefHandle                          crossHairRef;  // 28
+		std::uint32_t                            pad2C;         // 2C
+		TESQuest*                                quest;         // 30
+		TESWordOfPower*                          wordOfPower;   // 38
+		bool                                     show;          // 40
+		std::uint8_t                             pad41;         // 41
+		std::uint16_t                            pad42;         // 42
+		REX::EnumSet<MARKER_TYPE, std::uint32_t> discovery;     // 44
 	};
 	static_assert(sizeof(HUDData) == 0x48);
 }
