+++ include/RE/H/HighProcessData.h
@@ -362,7 +362,7 @@ namespace RE
 		bool                                                  fistsDrawn;                                // 46C
 		bool                                                  unk46D;                                    // 46D
 		bool                                                  unk46E;                                    // 46E
-		bool                                                  unk46F;                                    // 46F
+		bool                                                  allowRotation;                             // 46F
 		bool                                                  doorActivated;                             // 470
 		bool                                                  unk471;                                    // 471
 		bool                                                  aggroRadiusStarted;                        // 472
