+++ include/RE/M/MapMenu.h
@@ -55,7 +55,7 @@ namespace RE
 		BSTSmartPointer<MapZoomHandler> zoomHandler;           // 00050
 		ObjectRefHandle                 mapMarker;             // 00058
 		LocalMapMenu                    localMapMenu;          // 00060
-		RefHandle                       unk30460;              // 30460
+		RefHandle                       cameraOpeningCenter;   // 30460 - defaults to player ref
 		NiPoint3                        playerMarkerPosition;  // 30464
 		BSTArray<MapMenuMarker>         mapMarkers;            // 30470
 		BSTArray<GFxValue>              markerData;            // 30488
