+++ include/RE/T/TESQuest.h
@@ -13,6 +13,7 @@
 #include "RE/Q/QuestObjectiveStates.h"
 #include "RE/T/TESCondition.h"
 #include "RE/T/TESFullName.h"
+#include "RE/T/TeleportPath.h"
 
 namespace RE
 {
@@ -148,17 +149,16 @@ namespace RE
 			kCompassMarkerIgnoresLocks = 1 << 0
 		};
 
+		RefHandle& GetTargetReference(RefHandle& a_out, bool a_useExtraList, TESQuest* a_quest);
+
 		// members
-		std::uint64_t unk00;       // 00
-		TESCondition  conditions;  // 08
-		std::uint32_t alias;       // 10
-		std::uint32_t unk14;       // 14 - padding?
-		// These two fields might be a structure of some sort - in ae, the function
-		// w/ id 30980 frees both of these fields, taking a pointer to the first.
-		BSTArray<BSTempEffect*> unk18;  // 18
-		BSTArray<BSTempEffect*> unk30;  // 30
+		std::uint64_t unk00;         // 00
+		TESCondition  conditions;    // 08
+		std::uint32_t alias;         // 10
+		std::uint32_t unk14;         // 14
+		TeleportPath  teleportPath;  // 18
 	};
-	static_assert(sizeof(TESQuestTarget) == 0x48);
+	static_assert(sizeof(TESQuestTarget) == 0x60);
 
 	class BGSQuestObjective
 	{
