+++ src/RE/A/Actor.cpp
@@ -943,6 +943,12 @@ namespace RE
 		return boolFlags.all(BOOL_FLAGS::kProtected);
 	}
 
+	bool Actor::IsRotationAllowed() const
+	{
+		bool result = false;
+		return GetGraphVariableBool("bAllowRotation", result) && result;
+	}
+
 	bool Actor::IsRunning() const
 	{
 		using func_t = decltype(&Actor::IsRunning);
@@ -965,6 +971,15 @@ namespace RE
 		}
 
 		return true;
+	}
+
+	bool Actor::IsStaggering() const
+	{
+		bool result = false;
+		if (GetGraphVariableBool("IsStaggering", result) && result)
+			return result;
+
+		return ActorState::IsStaggered();
 	}
 
 	bool Actor::IsSummoned() const noexcept
